9.2.9. Store to mocks refactor
