+++ project/src/pages/main-empty-screen/main-empty-screen.test.tsx
@@ -2,33 +2,21 @@ import {Routes, Route} from 'react-router-dom';
 import {Provider} from 'react-redux';
 import {createMemoryHistory} from 'history';
 import {render, screen} from '@testing-library/react';
-import {configureMockStore} from '@jedmao/redux-mock-store';
 
 import HistoryRouter from '../../components/history-route/history-route';
 import MainEmptyScreen from './main-empty-screen';
 
-import {AppRoute, AuthorizationStatus, cities} from '../../const';
+import {AppRoute, cities} from '../../const';
 
-import {makeFakeOffers, makeFakeOffer} from '../../utils';
+import {store} from '../../utils';
 
-const mockOffers = makeFakeOffers();
-const mockOffer = makeFakeOffer();
-
-const mockStore = configureMockStore();
-const history = createMemoryHistory();
 const activeCity = 'Paris';
+const history = createMemoryHistory();
 
 global.window.scrollTo = jest.fn();
 
 describe('Component: MainEmptyScreen', () => {
   it('should render correctly', () => {
-    const store = mockStore({
-      USER: {authorizationStatus: AuthorizationStatus.NoAuth},
-      MAIN: {offers: mockOffers},
-      OFFER: {offer: mockOffer},
-      FAVORITE: {favorites: []},
-    });
-
     history.push(AppRoute.MainEmpty);
 
     render(
@@ -52,6 +40,4 @@ describe('Component: MainEmptyScreen', () => {
     expect(screen.getByText(/No places to stay available/i)).toBeInTheDocument();
     expect(screen.getByText(`We could not find any property available at the moment in ${activeCity}`)).toBeInTheDocument();
   });
-
 });
-
